Add auto-publish functionality for `Function` resources
This patch introduces the ability for CF Functions to be automatically
published after a successful create or update operation. It adds an
annotation `cloudfront.services.k8s.aws/auto-publish` to the `Function`
resource, which when set to `true`, will trigger the automatic
publishing of the Function after a create or update.

Signed-off-by: Amine Hilaly <hilalyamine@gmail.com>